requirements.txt
# Core dependencies (required)
pdf2image>=3.1.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Translation dependencies (choose one)
# Option 1: Google Cloud (paid, high quality)
google-cloud-translate>=3.12.0

# Option 2: Local translation (free, offline)
# Install ONE of these combinations:
# transformers>=4.21.0 torch>=1.12.0 sentencepiece>=0.1.95  # HuggingFace models (recommended)
# argostranslate>=1.8.0  # Fully offline (lightweight)
# requests>=2.28.0  # For LibreTranslate API access